pyproject.toml
[project]
name = "lima"
version = "0.1.0"
description = ""
authors = [
    {name = "Matheus Lima",email = "l1ma@live.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "alembic (>=1.15.2,<2.0.0)",
    "fastapi[standard] (>=0.115.12,<0.116.0)",
    "pwdlib[argon2] (>=0.2.1,<0.3.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "aiosqlite (>=0.21.0,<0.22.0)",
    "sqlalchemy[asyncio] (>=2.0.40,<3.0.0)",
    "psycopg[binary] (>=3.2.6,<4.0.0)",
    "tzdata (>=2025.2,<2026.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-cov = "^6.1.1"
taskipy = "^1.14.1"
ruff = "^0.11.7"
httpx = "^0.28.1"
pytest-asyncio = "^0.26.0"
factory-boy = "^3.3.3"
freezegun = "^1.5.1"
testcontainers = "^4.10.0"
pyngrok = "^7.2.5"

[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT', 'FAST']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'
asyncio_default_fixture_loop_scope = 'function'

[tool.coverage.run]
concurrency = ["thread", "greenlet"]

[tool.taskipy.tasks]
lint = 'ruff check'
pre_format = 'ruff check --fix'
format = 'ruff format'
run = 'fastapi dev lima/app.py'
pre_test = 'task lint'
test = 'pytest -s -x --cov=lima -vv'
post_test = 'coverage html'
# Tarefas para webhook e ngrok
server = 'uvicorn app:app --reload'
ngrok = 'ngrok http 8000'
webhook-config = 'python configure_webhook.py'
webhook-setup = { cmd = "echo 'Iniciando configuração do webhook do WhatsApp...' && task webhook-config", help = "Configura o webhook do WhatsApp" }
webhook-dev = { cmd = "python webhook_dev.py", help = "Inicia ambiente de desenvolvimento com ngrok e servidor" }

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"